requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.30.1
azure-cognitiveservices-speech==1.38.0
aiofiles==23.2.1
starlette==0.37.2
